# fluid-remote/docs/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: API for managing virtual machine sandboxes using libvirt
  title: fluid-remote API
  version: 0.0.1-beta
servers:
  - url: /
...
paths:
...
  /v1/ansible/jobs:
    post:
      description: Creates a new Ansible playbook execution job
      operationId: createAnsibleJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/fluid-remote_internal_ansible.JobRequest"
        description: Job creation parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_ansible.JobResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Bad Request
      summary: Create Ansible job
      tags:
        - Ansible
      x-codegen-request-body-name: request
  /v1/ansible/jobs/{job_id}:
    get:
      description: Gets the status of an Ansible job
      operationId: getAnsibleJob
      parameters:
        - description: Job ID
          in: path
          name: job_id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_ansible.Job"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Get Ansible job
      tags:
        - Ansible
  /v1/ansible/jobs/{job_id}/stream:
    get:
      description: Connects via WebSocket to run an Ansible job and stream output
      operationId: streamAnsibleJobOutput
      parameters:
        - description: Job ID
          in: path
          name: job_id
          required: true
          schema:
            type: string
      responses:
        "101":
          content:
            "*/*":
              schema:
                type: string
          description: Switching Protocols - WebSocket connection established
        "404":
          content:
            "*/*":
              schema:
                type: string
          description: Invalid job ID
        "409":
          content:
            "*/*":
              schema:
                type: string
          description: Job already started or finished
      summary: Stream Ansible job output
      tags:
        - Ansible
...
components:
  schemas:
...
    fluid-remote_internal_ansible.Job:
      example:
        check: true
        id: id
        vm_name: vm_name
        playbook: playbook
        status: pending
      properties:
        check:
          type: boolean
        id:
          type: string
        playbook:
          type: string
        status:
          $ref: "#/components/schemas/fluid-remote_internal_ansible.JobStatus"
        vm_name:
          type: string
      type: object
    fluid-remote_internal_ansible.JobRequest:
      properties:
        check:
          type: boolean
        playbook:
          type: string
        vm_name:
          type: string
      type: object
    fluid-remote_internal_ansible.JobResponse:
      example:
        job_id: job_id
        ws_url: ws_url
      properties:
        job_id:
          type: string
        ws_url:
          type: string
      type: object
    fluid-remote_internal_ansible.JobStatus:
      enum:
        - pending
        - running
        - finished
        - failed
      type: string
      x-enum-varnames:
        - JobStatusPending
        - JobStatusRunning
        - JobStatusFinished
        - JobStatusFailed
...
    fluid-remote_internal_error.ErrorResponse:
      example:
        code: 0
        details: details
        error: error
      properties:
        code:
          type: integer
        details:
          type: string
        error:
          type: string
      type: object
...